Admins can send emails to anyone now.
I wasn't able to fully test this, so please let me know if this works as I expect. I'm 70% certain it's good. (Download a mail software if u don't have one)
The Admin should be able to email anyone from the list of users by using a mail software on someone's computer.